+++ arctic/store/version_store.py
@@ -133,6 +133,9 @@ class VersionStore(object):
             query['symbol'] = {'$regex': regex}
         if kwargs:
             for k, v in six.iteritems(kwargs):
+                # TODO: this doesn't work as expected as it ignores the versions with metadata.deleted set
+                #       as a result it will return symbols with matching metadata which have been deleted
+                #       Maybe better add a match step in the pipeline instead of making it part of the query
                 query['metadata.' + k] = v
         if snapshot is not None:
             try:
@@ -148,22 +151,32 @@ class VersionStore(object):
             # Match based on user criteria first
             pipeline.append({'$match': query})
         pipeline.extend([
-                        # Id is by insert time which matches version order
-                        {'$sort': bson.SON([('symbol', pymongo.DESCENDING), ('version', pymongo.DESCENDING)])},
-                        # Group by 'symbol'
-                        {'$group': {'_id': '$symbol',
-                                    'deleted': {'$first': '$metadata.deleted'},
-                                    },
-                         },
-                        # Don't include symbols which are part of some snapshot, but really deleted...
-                        {'$match': {'deleted': {'$ne': True}}},
-                        {'$project': {'_id': 0,
-                                      'symbol':  '$_id',
-                                      }
-                         }])
+            # version_custom value is: 2*version + (0 if deleted else 1)
+            # This is used to optimize aggregation query:
+            #  - avoid sorting
+            #  - be able to rely on the latest version (max) for the deleted status
+            #
+            # Be aware of that if you don't use custom sort or if use a sort before $group which utilizes
+            # exactly an existing index, the $group will do best effort to utilize this index:
+            #  - https://jira.mongodb.org/browse/SERVER-4507
+            {'$group': {
+                '_id': '$symbol',
+                'version_custom': {
+                    '$max': {
+                        '$add': [
+                            {'$multiply': ['$version', 2]},
+                            {'$cond': [{'$eq': ['$metadata.deleted', True]}, 1, 0]}
+                        ]
+                    }
+                },
+            }},
+            # Don't include symbols which are part of some snapshot, but really deleted...
+            {'$match': {'version_custom': {'$mod': [2, 0]}}}
+        ])
 
-        results = self._versions.aggregate(pipeline)
-        return sorted([x['symbol'] for x in results])
+        # We may hit the group memory limit (100MB), so use allowDiskUse to circumvent this
+        #  - https://docs.mongodb.com/manual/reference/operator/aggregation/group/#group-memory-limit
+        return sorted([x['_id'] for x in self._versions.aggregate(pipeline, allowDiskUse=True)])
 
     @mongo_retry
     def has_symbol(self, symbol, as_of=None):