+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### 1.62
   * Bugfix: #517 VersionStore: append does not duplicate data in certain corner cases
+  * Bugfix: #519 VersionStore: list_symbols speed improvement and fix for memory limit exceed
 
 ### 1.61 (2018-3-2)
   * Feature: #288 Mapping reads and writes over chunks in chunkstore